consumer_service/requirements.txt
kafka-python
minio
